system-python-support-for-requirements

+++ local-gen/installer.py
@@ -213,7 +213,10 @@ def install_comfyui_nodes(custom_nodes_folder : str) -> None:
 		req_txtfile = os.path.join(custom_nodes_folder, folder_name, "requirements.txt")
 		if os.path.exists(req_txtfile):
 			print(f'Installing requirements for: {folder_name} {req_txtfile}')
-			run_command(f"{py_exe} -m pip install -r {Path(req_txtfile).as_posix()}")
+			if os.path.exists(py_exe):
+				run_command(f"{py_exe} -m pip install -r {Path(req_txtfile).as_posix()}")
+			else:
+				run_command(f"{PYTHON_COMMAND} -m pip install -r {Path(req_txtfile).as_posix()}")
 	print("Installed ComfyUI Custom Nodes")
 
 def prompt_safetensor_file_install(folder : str, filename : str, download_url : str) -> None:
